fix: keep the latency's decimals instead of rounding them away
The ping column was declared INTEGER, so every measurement was rounded to
whole milliseconds before it was stored - and sequelize validates against
the declared type, so a fractional value was refused outright rather than
merely truncated. On a fibre or a local line most of the reading lives
below the millisecond: an idle latency of 0.4 ms and one of 1.4 ms were
the same row. The rounding is not recoverable afterwards, because the API,
the CSV export, the Prometheus exporter and every integration read this
column and nothing else. Jitter, packet loss and both loaded latencies
beside it have always been DOUBLE; the one figure people actually watch
was the exception.

The three aggregates that rounded to whole milliseconds - the min/max/avg
tiles, the hourly buckets and the downsampled chart series - now hold two
decimals like every metric beside them, so the precision is not discarded
at the last step instead of the first.

The migration leaves sqlite alone, for the same reason 0007 does:
sequelize rebuilds the table there from describeTable's output, which does
not report autoIncrement, so the rebuild strips AUTOINCREMENT off the
primary key and takes the created index with it. Nothing is lost by
skipping it - sqlite's INTEGER affinity is numeric, so a fraction round
trips there already. MySQL's INT is the only one that actually rounds.

The interface shows one decimal, which is what the issue asks for: the
stored value and the API keep two, but the test row is dense and the
second decimal is noise at a glance. Existing rows keep their whole
millisecond values and gain decimals from the next test.

Upstream gnmyt/MySpeed#1387 and #999.

+++ server/models/Speedtests.js
@@ -30,8 +30,12 @@ export default db.define("speedtests", {
         allowNull: true,
         defaultValue: null
     },
+    // DOUBLE, like every other measurement on the row. As INTEGER this rounded
+    // the latency to whole milliseconds before it was stored - which is most of
+    // the reading on a fibre or local line, and unrecoverable afterwards since
+    // the API, the export and the integrations all read the stored column.
     ping: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
         allowNull: false
     },
     jitter: {
